renaming MayoLink Api class as MayoLink Client and moving to services directory
